Refactors based on Jeremy's comments
