+++ packages/front-end/components/Features/FeaturesOverview.tsx
@@ -73,7 +73,6 @@ import Badge from "@/components/Radix/Badge";
 import Frame from "@/components/Radix/Frame";
 import LoadingSpinner from "@/components/LoadingSpinner";
 import JSONValidation from "@/components/Features/JSONValidation";
-import PremiumCallout from "../Radix/PremiumCallout";
 import PrerequisiteStatusRow, {
   PrerequisiteStatesCols,
 } from "./PrerequisiteStatusRow";
@@ -539,18 +538,6 @@ export default function FeaturesOverview({
         <Heading mb="3" size="5" as="h2">
           Overview
         </Heading>
-
-        <PremiumCallout
-          id="ff-dev-tools"
-          commercialFeature={null}
-          dismissable={true}
-          docSection="devTools"
-          mb="3"
-        >
-          Download our [Browser] DevTools extension to easily preview and QA
-          feature flags in your application.
-        </PremiumCallout>
-        {/* <Callout status="wizard">Hi</Callout> */}
 
         <Frame>
           <div className="mh-350px" style={{ overflowY: "auto" }}>
